Bootstrap sgcca

Function to perform bootstrap on the samples. boot_samples_sgcca bootstrap given original data, while boot_index_sgcca given some index of samples it iterates over it.

boot_samples_sgcca(..., nb_boot = 1000, verbose = TRUE)

boot_index_sgcca(index, ...)

Arguments

...
Named arguments passed to sgcca.

nb_boot
Number of bootstraps to perform.

verbose
Logical, should it print a progress bar (default) or not?

index
A list of numeric values for selecting values

Value

A list with two elements: the coefficient of each variable of the input blocks; and the AVE values, both inner, and outer

Details

boot_index_sgcca Iterate over the index, which is a list of vectors with the position of samples to use. boot_samples_sgcca Iterate over random samples.

Examples

data("Russett", package = "RGCCA")
X_agric <- as.matrix(Russett[, c("gini", "farm", "rent")])
X_ind <- as.matrix(Russett[, c("gnpr", "labo")])
X_polit <- as.matrix(Russett[ , c("inst", "ecks",  "death", "demostab",
                                  "dictator")])
A <- list(X_agric, X_ind, X_polit)
C <- matrix(c(0, 0, 1, 0, 0, 1, 1, 1, 0), 3, 3)
out <- boot_samples_sgcca(A = A, C = C, c1 = rep(1, 3),  nb_boot = 10)

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

head(out$AVE)

#>          inner     outer
#> [1,] 0.4037123 0.6377082
#> [2,] 0.4141027 0.7195481
#> [3,] 0.4127596 0.6844097
#> [4,] 0.4755541 0.6865155
#> [5,] 0.3604672 0.6768108
#> [6,] 0.3914483 0.6850278

boots <- 10
index <- vector("list", length = boots)
for (i in seq_len(boots)) {
  index[[i]] <- sample(nrow(A[[1]]), replace = TRUE)
}
boot_i <- boot_index_sgcca(index, A = A, C = C)